# app/Http/Controllers/Api/V3/PostController.php
<?php

namespace App\Http\Controllers\Api\V3;

use App\Repositories\Contracts\PostRepository;
use App\Transformers\PostTransformer;
use Illuminate\Http\Request;
use Swagger\Annotations as SWG;

class PostController extends BaseController
{
    private $postRepository;

    public function __construct(PostRepository $postRepository)
    {
        $this->postRepository = $postRepository;
    }

    /**
     * @SWG\Get(
     *   path="/posts",
     *   tags={"Post"},
     *   summary="帖子列表(post list)",
     *   description="帖子列表(post list)",
     *   operationId="index",
     *   consumes={"application/json"},
     *   produces={"application/json"},
     *   @SWG\Parameter(
     *     name="include",
     *     in="query",
     *     type="string",
     *     description="comments:limit(x), user",
     *   ),
     *
     *   @SWG\Response(
     *     response=200,
     *     description="分页的帖子列表",
     *     @SWG\Schema(
     *       type="array",
     *       @SWG\Items(ref="#/definitions/Post")
     *     ),
     *   )
     * )
     */
    public function index()
    {
        $posts = $this->postRepository->paginate();

        return $this->response->paginator($posts, new PostTransformer());
    }

    /**
     * @SWG\Get(
     *   path="/user/posts",
     *   description="我的帖子列表(my post list)",
     *   summary="我的帖子列表(my post list)",
     *   tags={"Post"},
     *   operationId="userIndex",
     *   consumes={"application/json"},
     *   produces={"application/json"},
     *   @SWG\Parameter(
     *     name="include",
     *     type="string",
     *     description="comments:limit(x)",
     *     required=false,
     *     in="query"
     *   ),
     *   @SWG\Response(
     *     response=200,
     *     description="我的帖子列表",
     *     @SWG\Schema(
     *       type="object",
     *       @SWG\Property(
     *          property="data",
     *          type="array",
     *          @SWG\Items(
     *            ref="#/definitions/Post"
     *          )
     *       ),
     *       @SWG\Property(
     *         property="meta",
     *         type="object",
     *         @SWG\Property(
     *           property="pagination",
     *           type="object",
     *           @SWG\Property(
     *             property="total",
     *             type="int",
     *             default="50"
     *           ),
     *           @SWG\Property(
     *             property="count",
     *             type="int",
     *             default="15",
     *           ),
     *           @SWG\Property(
     *             property="per_page",
     *             type="int",
     *             default="15",
     *           ),
     *           @SWG\Property(
     *              property="current_page",
     *              type="int",
     *              default="1"
     *           ),
     *           @SWG\Property(
     *              property="total_pages",
     *              type="int",
     *              default="4"
     *           ),
     *           @SWG\Property(
     *             property="links",
     *             type="object",
     *             @SWG\Property(
     *                property="next",
     *                type="string",
     *                default="http://localhost:8800/api/posts?page=2"
     *             ),
     *              @SWG\Property(
     *                  property="prev",
     *                  type="string",
     *                  default="http://localhost:8800/api/posts?page=2"
     *              )
     *           )
     *         )
     *       )
     *     )
     *   ),
     *   security={{
     *     "jwt":{}
     *   }}
     * )
     *
     */
    public function userIndex()
    {
        $posts = $this->postRepository
            ->where(['user_id' => $this->user()->id])
            ->paginate();

        return $this->response->paginator($posts, new PostTransformer());
    }

    /**
     * @api            {get} /posts/{id} 帖子详情(post detail)
     * @apiDescription 帖子详情(post detail)
     * @apiGroup       Post
     * @apiPermission  none
     * @apiParam {String='comments','user'} [include]  include
     * @apiVersion     0.1.0
     * @apiSuccessExample {json} Success-Response:
     *   HTTP/1.1 200 OK
     *   {
     *     "data": {
     *       "id": 1,
     *       "user_id": 3,
     *       "title": "foo",
     *       "content": "",
     *       "created_at": "2016-03-30 15:36:30",
     *       "user": {
     *         "data": {
     *           "id": 3,
     *           "email": "foo@bar.com1",
     *           "name": "",
     *           "avatar": "",
     *           "created_at": "2016-03-30 15:34:01",
     *           "updated_at": "2016-03-30 15:34:01",
     *           "deleted_at": null
     *         }
     *       },
     *       "comments": {
     *         "data": [
     *           {
     *             "id": 1,
     *             "post_id": 1,
     *             "user_id": 1,
     *             "reply_user_id": 0,
     *             "content": "foobar",
     *             "created_at": "2016-04-06 14:51:34"
     *           }
     *         ],
     *         "meta": {
     *           "total": 1
     *         }
     *       }
     *     }
     *   }
     */
    public function show($id)
    {
        $post = $this->postRepository->find($id);

        if (!$post) {
            return $this->response->errorNotFound();
        }

        return $this->response->item($post, new PostTransformer());
    }

    /**
     * @api            {post} /posts 发布帖子(create post)
     * @apiDescription 发布帖子(create post)
     * @apiGroup       Post
     * @apiPermission  jwt
     * @apiParam {String} title  post title
     * @apiParam {String} content  post content
     * @apiVersion     0.1.0
     * @apiSuccessExample {json} Success-Response:
     * HTTP/1.1 201 Created
     * {
     *     "data": {
     *         "title": "foo",
     *         "content": "bar",
     *         "user_id": 33,
     *         "created_at": "2017-01-19 15:37:25",
     *         "id": 212
     *     }
     * }
     */
    public function store(Request $request)
    {
        $validator = \Validator::make($request->input(), [
            'title'   => 'required|string|max:50',
            'content' => 'required|string',
        ]);

        if ($validator->fails()) {
            return $this->errorBadRequest($validator);
        }

        $attributes = $request->only('title', 'content');
        $attributes['user_id'] = $this->user()->id;
        $post = $this->postRepository->create($attributes);

        $location = dingo_route('v1', 'posts.show', $post->id);
        // 不返回数据
        // return $this->response->created($location);

        // 返回数据
        return $this->response
            ->item($post, new PostTransformer())
            ->withHeader('Location', $location)
            ->setStatusCode(201);
    }

    /**
     * @api            {put} /posts/{id} 替换帖子(update post)
     * @apiDescription 替换帖子(update post)
     * @apiGroup       Post
     * @apiPermission  jwt
     * @apiParam {String} title  post title
     * @apiParam {String} content  post content
     * @apiVersion     0.1.0
     * @apiSuccessExample {json} Success-Response:
     *   HTTP/1.1 204 No Content
     */
    public function update($id, Request $request)
    {
        $post = $this->postRepository->find($id);

        if (!$post) {
            return $this->response->errorNotFound();
        }

        // 不属于我的forbidden
        if ($post->user_id != $this->user()->id) {
            return $this->response->errorForbidden();
        }

        $validator = \Validator::make($request->input(), [
            'title'   => 'required|string|max:50',
            'content' => 'required|string',
        ]);

        if ($validator->fails()) {
            return $this->errorBadRequest($validator);
        }

        $this->postRepository->update($id, $request->only('title', 'content'));

        return $this->response->noContent();
    }

    /**
     * @api            {patch} /posts/{id} 修改帖子(update part of post)
     * @apiDescription 修改帖子(update part of post)
     * @apiGroup       Post
     * @apiPermission  jwt
     * @apiParam {String} [title]  post title
     * @apiParam {String} [content]  post content
     * @apiVersion     0.1.0
     * @apiSuccessExample {json} Success-Response:
     *   HTTP/1.1 204 No Content
     */
    public function patch($id, Request $request)
    {
        $post = $this->postRepository->find($id);

        if (!$post) {
            return $this->response->errorNotFound();
        }

        // 不属于我的forbidden
        if ($post->user_id != $this->user()->id) {
            return $this->response->errorForbidden();
        }

        $validator = \Validator::make($request->input(), [
            'title'   => 'string|max:50',
            'content' => 'string',
        ]);

        if ($validator->fails()) {
            return $this->errorBadRequest($validator);
        }

        $this->postRepository->update(
            $id,
            array_filter($request->only('title', 'content'))
        );

        return $this->response->noContent();
    }

    /**
     * @api            {delete} /posts/{id} 删除帖子(delete post)
     * @apiDescription 删除帖子(delete post)
     * @apiGroup       Post
     * @apiPermission  jwt
     * @apiVersion     0.1.0
     * @apiSuccessExample {json} Success-Response:
     *   HTTP/1.1 204 NO CONTENT
     */
    public function destroy($id)
    {
        $post = $this->postRepository->find($id);

        if (!$post) {
            return $this->response->errorNotFound();
        }

        // 不属于我的forbidden
        if ($post->user_id != $this->user()->id) {
            return $this->response->errorForbidden();
        }

        $this->postRepository->destroy($id);

        return $this->response->noContent();
    }
}